# bytes, strings: speed up TrimSpace